Created Solution in Visual Studio 2013 Professional
Please use the files stored in the "Black Jack Game" directory. Those
not in that directory are now irrelevant and won't be kept up to date.

Now we need to finish the linkedlist, node, player, and hand classes
(maybe others, everything needs to be tested/checked over).

Best of luck to us!